Print paths for unmounted intermediate datasets
Fixes #5

This finds the longest common prefix of path segments since the last and
current dataset and prints any up to the final segment that have been
missed.

For example, a root mount at rpool/ROOT/default would previously be
displayed thus:

dataset
        default ...

This ensures the path leading up to it have their own lines:

dataset
rpool
    ROOT
        default ...

I believe this to be a more consistent fix than the one proposed in #6
where intermediates are formatted on a single line.