Add -pP flags and rework sorting
This decouples the tree-style dataset path display from the sort mode,
allowing the user to override it with -p and -P flags, while having the
default behavior of only using the tree-style abbreviated approach for
name sorting.

Sorts are moved to a dict and the keys used to configure argparse.

Name sorting is now performed unconditionally so name is used as a
secondary sort key to a primary sort.