Restore default SIGINT and SIGPIPE handlers
Instead of catching SIGINT by way of a Python exception handler and
exiting normally, just let the system default handler terminate us.

Do the same for SIGPIPE, rather than messing about with a handler for
BrokenPipeError.  This is discouraged by Python docs, but we're only
writing to stdout/err.

This fixes an uncaught BrokenPipeError, and ensures our signaled status
propagates to the caller via WIFSIGNALED().