Linux: Fix possible FileNotFoundError on dataset destroy
Since /proc is a silly design unsuited to point-in-time aggregate kernel
statistics, a race condition exists between enumerating the files we're
interested in and opening them.

If a dataset is destroyed in between our glob() and open() calls, the
respective file will disappear out from under us. Swallow such
exceptions.