FreeBSD: Fix CalledProcessError on non-existant pools
With sysctl(8) fallback, a non-existant pool will raise an
unhandled subprocess.CalledProcessError exception due to a non-zero
exit code.

For the moment the other two data gathering methods, py-sysctl and
/proc, silently ignore missing pools.  Make the behaviour consistant by
setting check=False.

This error - if we wish to handle it - should probably be done so
generically by detecting if any requested pools are missing in the
initial summary.