Remove regexp use for dataset matching
Instead of building a big regexp out of every specified dataset with
anchors based on recursive mode, use x in y and x.startswith z.

Not only is this tidier, it's also faster.

While here, be more defensive about the dataset names we accept - in
particular, reject empty strings, and reject trailing slashes, both of
which would otherwise cause problems.