Add unlink statistics behind an -x flag
Add the two other fields ZFS exposes: nunlinked, the number of files
unlinked from the filesystem, and nunlinks, which includes the number
that have been queued.

Add an -x flag for extended statistics.  Hide both opsize and nunlinks
by default to make for a clear display, with -x to add opsize and -xx to
add both that and unlink statistics.