Refactor create_diff_iter into a class
Reduce all the 'diff_loop = ... for diffs in diff_loop' noise by
encapsulating the iterator in an object with methods for chaining
operations on the iterator.

I had to add an explicit list() to the end of the iterator chain which
wasn't there before.  Looking at the code, I'm actually unclear as to
how it worked without this, as if we ended up with an iterator of
generators we'd consume it when calculating the max name width.