Correct diff calculation
Make Dataset.per_second() produce a copy of the Dataset with normalized
values instead of modifying it in-place, as this breaks diff
calculation.

Also correctly handle timestamps - this was producing ostensibly correct
values only because per_second was no-oping as the timespan was zero.

Remove a dead comment while here.