Make fetch_datasets() take a list of pools
Instead of calling fetch_datasets individually for each pool, pass it an
optional list.  These functions must already be able to fetch multiple
pools as their default behaviour is to fetch all of them.

This removes the need for dataset_iter, as it simply reduces to calling
the function repeatedly, reduces the amount of temporary objects that
need creating, and with some more work could make FreeBSD's sysctl(8)
fallback more efficient.